add a more clever solution to day 19 problem (part 1)
it expands the leftmost rules. compares the leftmost rule with the
leftmost entry rule. If both are terminal and matches, keep expanding
and searching the rest of the entry.

if there are rules and entry is empty or if there are no rules and
entry isn't empty it means that is not a match. Otherwise it is since
both were consumed